Create SQL trigger to increment user_votes
Message creation is now responsible for augmenting each message entry
with the correct userToken. This is not a problem since userToken is
freely avalible on userData in the models anyway.

An SQL trigger is declared in the schema which automatically increments
a users votes whenever one of their messages in upvoted. This works
assuming each message the user has includes the userToken field.

userToken is not included as a foreign key in the messages table to make
deletion more simple.